Add base entity and feat
The base entity is the root class for all entities in the type
hierarchy which could be imported. This basically includes everything
except attributes and skills, which are hard-coded into the library.

The feat class represents a single feat, adding a prerequistes element
as well as a list of feat types.